Refactor resource router
